+++ app/controllers/v1/application_controller.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+
+module V1
+  class ApplicationController < ActionController::API
+  end
+end
