+++ config/routes.rb
@@ -1,5 +1,7 @@
 # frozen_string_literal: true
 
 Rails.application.routes.draw do
-  resources :notes
+  namespace :v1 do
+    resources :notes
+  end
 end
