--- spec/routing/notes_routing_spec.rb
@@ -1,31 +1,0 @@
-# frozen_string_literal: true
-
-require 'rails_helper'
-
-RSpec.describe NotesController, type: :routing do
-  describe 'routing' do
-    it 'routes to #index' do
-      expect(get: '/notes').to route_to('notes#index')
-    end
-
-    it 'routes to #show' do
-      expect(get: '/notes/1').to route_to('notes#show', id: '1')
-    end
-
-    it 'routes to #create' do
-      expect(post: '/notes').to route_to('notes#create')
-    end
-
-    it 'routes to #update via PUT' do
-      expect(put: '/notes/1').to route_to('notes#update', id: '1')
-    end
-
-    it 'routes to #update via PATCH' do
-      expect(patch: '/notes/1').to route_to('notes#update', id: '1')
-    end
-
-    it 'routes to #destroy' do
-      expect(delete: '/notes/1').to route_to('notes#destroy', id: '1')
-    end
-  end
-end
