+++ spec/routing/v1/notes_routing_spec.rb
@@ -1,0 +1,31 @@
+# frozen_string_literal: true
+
+require 'rails_helper'
+
+RSpec.describe V1::NotesController, type: :routing do
+  describe 'routing' do
+    it 'routes to #index' do
+      expect(get: '/v1/notes').to route_to('v1/notes#index')
+    end
+
+    it 'routes to #show' do
+      expect(get: '/v1/notes/1').to route_to('v1/notes#show', id: '1')
+    end
+
+    it 'routes to #create' do
+      expect(post: '/v1/notes').to route_to('v1/notes#create')
+    end
+
+    it 'routes to #update via PUT' do
+      expect(put: '/v1/notes/1').to route_to('v1/notes#update', id: '1')
+    end
+
+    it 'routes to #update via PATCH' do
+      expect(patch: '/v1/notes/1').to route_to('v1/notes#update', id: '1')
+    end
+
+    it 'routes to #destroy' do
+      expect(delete: '/v1/notes/1').to route_to('v1/notes#destroy', id: '1')
+    end
+  end
+end
